현구님 코멘트 반영
