+++ src/main/kotlin/racingcar/domain/RandomUtil.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 import java.util.concurrent.ThreadLocalRandom
