+++ src/main/kotlin/racingcar/input/InputView.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.input
 
 object InputView {
