+++ src/main/kotlin/racingcar/output/ResultView.kt
@@ -1,9 +1,12 @@
-package racingcar
+package racingcar.output
+
+import racingcar.domain.Car
+import racingcar.domain.Position
 
 object ResultView {
 
     fun showCurrentRound(round: Int){
-      println("현재 라운드는 $round 회차 입니다.")
+      println("현재 라운드는 ${round+1} 회차 입니다.")
     }
 
     fun showResult(cars: List<Car>) {
