+++ src/test/kotlin/racingcar/CarNumberTest.kt
@@ -8,6 +8,7 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
 import org.junit.jupiter.params.provider.ValueSource
+import racingcar.domain.vo.CarNumber
 import java.util.stream.Stream
 
 class CarNumberTest: AnnotationSpec(){
@@ -25,7 +26,7 @@ class CarNumberTest: AnnotationSpec(){
     fun `자동차의 개수로 정수가 아닌 문자열이 들어오면 안된다`(value: String){
         val exception = shouldThrow<IllegalArgumentException> { CarNumber(value) }
 
-        exception.message shouldBe "문자열이 아닌 정수값을 입력해주세요. 예시) 5"
+        exception.message shouldBe "문자열이 아닌 정수값이여야 합니다"
     }
 
     @ParameterizedTest
@@ -33,7 +34,7 @@ class CarNumberTest: AnnotationSpec(){
     fun `자동차의 개수로 0, 음수가 들어오면 안된다`(value: String){
         val exception = shouldThrow<IllegalArgumentException> { CarNumber(value) }
 
-        exception.message shouldBe "1이상의 양의 정수를 입력해주세요."
+        exception.message shouldBe "1이상의 양의 정수여야 합니다."
     }
 
     companion object {
