+++ src/test/kotlin/racingcar/RacingCarsTest.kt
@@ -3,6 +3,8 @@ package racingcar
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.inspectors.forAll
 import io.kotest.matchers.shouldBe
+import racingcar.domain.*
+import racingcar.domain.vo.CarNumber
 
 class RacingCarsTest: StringSpec({
